requirements.txt
discord.py
aiohttp
aiofiles
python-dotenv
PyGithub